Moved Leases class in sys package and deprecated old methods
